Fixed failing tests
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@50281 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ rm-server/test/java/org/alfresco/module/org_alfresco_module_rm/test/util/BaseRMWebScriptTestCase.java
@@ -353,7 +353,10 @@ public class BaseRMWebScriptTestCase extends BaseWebScriptTest
 
     protected void deleteUser(String userName)
     {
-        personService.deletePerson(userName);
+        if (authenticationService.authenticationExists(userName) == true)
+        {
+            personService.deletePerson(userName);
+        }
     }
 
     protected void createGroup(String groupName)
@@ -366,6 +369,9 @@ public class BaseRMWebScriptTestCase extends BaseWebScriptTest
 
     protected void deleteGroup(String groupName)
     {
-        authorityService.deleteAuthority(groupName, true);
+        if (authorityService.authorityExists(groupName) == true)
+        {
+            authorityService.deleteAuthority(groupName, true);
+        }
     }
 }
